refactor experiment logic into src/ package
Move GoEmotions prep, pipeline build/fit/report, and output-path/collision
helpers into src/{data,pipeline,naming}.py; prepare_data.py and run.py become
thin argparse wrappers. Add pre-launch collision checks with --overwrite and an
--exp-name to avoid clobbering outputs. Drop run.bash (orchestration now in Python).
